requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
dill==0.3.8
docopt==0.6.2
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
frozenlist==1.6.0
future==1.0.0
greenlet==3.2.1
huggingface-hub==0.30.2
HyperPyYAML==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
openai-whisper==20240930
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
preshed==3.0.9
primePy==1.3
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.40
srsly==2.5.1
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
